NA handling in proxy objects; clean up st_warp

+++ R/warp.R
@@ -104,7 +104,7 @@ transform_grid_grid = function(x, target) {
 #' @details For gridded spatial data (dimensions \code{x} and \code{y}), see figure; the existing grid is transformed into a regular grid defined by \code{dest}, possibly in a new coordinate reference system. If \code{dest} is not specified, but \code{crs} is, the procedure used to choose a target grid is similar to that of \link[raster]{projectRaster} (currently only with \code{method='ngb'}). This entails: (i) the envelope (bounding box polygon) is transformed into the new crs, possibly after segmentation (red box); (ii) a grid is formed in this new crs, touching the transformed envelope on its East and North side, with (if cellsize is not given) a cellsize similar to the cell size of \code{src}, with an extent that at least covers \code{x}; (iii) for each cell center of this new grid, the matching grid cell of \code{x} is used; if there is no match, an \code{NA} value is used.
 #' @export
 st_warp = function(src, dest, ..., crs = NA_crs_, cellsize = NA_real_, segments = 100, 
-		use_gdal = FALSE, options = character(0), no_data_value = -9999, debug = FALSE,
+		use_gdal = FALSE, options = character(0), no_data_value = NA_real_, debug = FALSE,
 		method = "near") {
 
 	if (!is.na(crs))
@@ -113,8 +113,11 @@ st_warp = function(src, dest, ..., crs = NA_crs_, cellsize = NA_real_, segments 
 	if (use_gdal) {
 		options = c(options, "-dstnodata", no_data_value, "-r", method)
 		if (! inherits(src, "stars_proxy")) {
-			src = st_as_stars_proxy(src)
-			on.exit(unlink(src[[1]])) # a temp file
+			src = st_as_stars_proxy(src, NA_value = no_data_value)
+			if (debug)
+				cat("Writing input to: ", src[[1]], "\n")
+			else
+				on.exit(unlink(src[[1]])) # temp file
 		}
 		if (missing(dest) && is.na(crs)) {
 			delete = TRUE
@@ -122,8 +125,9 @@ st_warp = function(src, dest, ..., crs = NA_crs_, cellsize = NA_real_, segments 
 			sf::gdal_utils("warp", src[[1]], dest, options = options)
 		} else {  # dest exists, and should be used: should use warper rather than warp
 			dest = if (! inherits(dest, "stars_proxy")) {
-					delete = TRUE
-					st_as_stars_proxy(dest)[[1]]
+					dest[[1]] = NA_real_ * dest[[1]] # blank out values
+					delete = !debug
+					st_as_stars_proxy(dest, NA_value = no_data_value)[[1]]
 				} else {
 					delete = FALSE
 					dest[[1]] # the file name of the stars_proxy, not to be deleted
@@ -134,8 +138,8 @@ st_warp = function(src, dest, ..., crs = NA_crs_, cellsize = NA_real_, segments 
 				sf::gdal_utils("warper", src[[1]], dest, options = method) # not "warp"!
 		}
 		if (debug)
-			cat("Writing to: ", dest, "\n")
-		if (delete)
+			cat("Writing result to: ", dest, "\n")
+		else if (delete)
 			on.exit(unlink(dest)) # a temp file
 		read_stars(dest)
 	} else {
